Upgrade to vitest 1.3.0.
That version deprecates an overload of `it` and introduces a new
one, where `testOptions` is the second argument. In order to
upgrade to future versions (compatible with newer Vite versions),
we need the new test code, and hence the new Vitest version.

Vitest 1.3.0 nominally requires Vite 5.x. However, empirically it
can cope with our Vite 4.x, so we keep it there.

It does not support Node.js v16, though. We upgrade to Node.js v18
instead.

Co-authored-by: NavidJalali <nvd.eth0@gmail.com>